chore: replace existing tooltip with newest implementation
