--- src/components/NewTooltip/NewTooltip.scss
@@ -1,56 +1,0 @@
-// Tooltip Container
-.deriv-tooltip-v2 {
-    z-index: 9999;
-    padding: 8px;
-    border-radius: $border-radius;
-    width: max-content;
-    max-width: 280px;
-    animation: fadeIn ease-in-out 0.2s;
-    font-size: 12px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 18px;
-    background-color: var(--du-general-active, #d6dadb);
-    color: var(--du-text-general, #333);
-}
-
-// Tooltip Arrow
-.deriv-tooltip-v2__arrow,
-.deriv-tooltip-v2__arrow::before {
-    position: absolute;
-    width: 8px;
-    height: 8px;
-    background: inherit;
-}
-
-.deriv-tooltip-v2__arrow {
-    visibility: hidden;
-}
-
-.deriv-tooltip-v2__arrow::before {
-    visibility: visible;
-    content: "";
-    transform: rotate(45deg);
-}
-
-// Tooltip Arrow Position
-.deriv-tooltip-v2[data-popper-placement^="top"] > .deriv-tooltip-v2__arrow {
-    bottom: -4px;
-}
-
-.deriv-tooltip-v2[data-popper-placement^="bottom"] > .deriv-tooltip-v2__arrow {
-    top: -4px;
-}
-
-.deriv-tooltip-v2[data-popper-placement^="left"] > .deriv-tooltip-v2__arrow {
-    right: -4px;
-}
-
-.deriv-tooltip-v2[data-popper-placement^="right"] > .deriv-tooltip-v2__arrow {
-    left: -4px;
-}
-
-// Tooltip Trigger Class
-.deriv-tooltip-v2__trigger {
-    cursor: pointer;
-}
